Split base data classes into their own PathLanguage module.
